// types/theme.d.ts
type ObjectOf<T> = { [key: string]: T }

// ========================================================
// Props
// ========================================================

export type IProps = ObjectOf<any>

// ========================================================
// Variables
// ========================================================

export interface ISiteVariables {
  [key: string]: any

  brand: string
  htmlFontSize: string
}

export type ComponentVariableValue = any

export type ComponentVariablesObject = ObjectOf<ComponentVariableValue>

export type ComponentVariablesFunction = (siteVariables: ISiteVariables) => ComponentVariablesObject

export type ComponentVariables = ComponentVariablesObject | ComponentVariablesFunction

// ========================================================
// Styles
// ========================================================

export interface ICSSPseudoElementStyle extends ICSSInJSStyle {
  content?: string
}

export interface ICSSInJSStyle extends React.CSSProperties {
  // missing React.CSSProperties
  speak?: CSSType.Globals | 'none' | 'normal' | 'spell-out'

  // CSS in JS Properties
  '::before'?: ICSSPseudoElementStyle
  '::after'?: ICSSPseudoElementStyle

  ':hover'?: ICSSInJSStyle
  ':active'?: ICSSInJSStyle
  ':focus'?: ICSSInJSStyle
  ':visited'?: ICSSInJSStyle

  ':first-child'?: ICSSInJSStyle
  ':last-child'?: ICSSInJSStyle
}

export type ComponentPartStyle = ComponentStyleFunction | ICSSInJSStyle

export interface IComponentStyles {
  [part: string]: ComponentPartStyle

  root?: ComponentPartStyle
}

export interface IComponentStyleClasses {
  [part: string]: string

  root?: string
}

export interface ComponentStyleFunctionArg {
  props: IProps
  variables: ComponentVariablesObject
  siteVariables: ISiteVariables
  rtl: boolean
}

export type ComponentStyleFunction = (arg: ComponentStyleFunctionArg) => ICSSInJSStyle